chore: finalize review-driven fixes

+++ KimiDiscordServer.Bot/Configuration/DiscordOptions.cs
@@ -12,7 +12,7 @@ public sealed class DiscordOptions
 
     public List<ulong> AllowedChannelIds { get; set; } = [];
 
-    public int MaxAttachmentBytes { get; set; } = 4 * 1024 * 1024;
+    public long MaxAttachmentBytes { get; set; } = 4 * 1024 * 1024;
 
     public int MaxTextDocumentBytes { get; set; } = 256 * 1024;
